chore: update utils with is_vue_next_tick_import
